Rename Package name & Fix Bottom Nav
- 패키지 이름 수정
- 바텀네비게이션 메인액티비티로 이동
