+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     compileSdkVersion compile_sdk_version
 
     defaultConfig {
-        applicationId "com.anonymous.coffin"
+        applicationId "com.anonymous.appilogue"
         minSdkVersion min_sdk_version
         targetSdkVersion target_sdk_version
         versionCode 1
