+++ app/src/androidTest/java/com/anonymous/appilogue/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.anonymous.coffin
+package com.anonymous.appilogue
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.anonymous.coffin", appContext.packageName)
+        assertEquals("com.anonymous.appilogue", appContext.packageName)
     }
 }
