+++ app/src/main/java/com/anonymous/appilogue/AnonymousApplication.kt
@@ -1,4 +1,4 @@
-package com.anonymous.coffin
+package com.anonymous.appilogue
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
