+++ app/src/main/java/com/anonymous/appilogue/di/RepositoryModule.kt
@@ -1,4 +1,4 @@
-package com.anonymous.coffin.di
+package com.anonymous.appilogue.di
 
 import dagger.Module
 import dagger.hilt.InstallIn
