+++ app/src/main/java/com/anonymous/appilogue/features/base/BaseFragment.kt
@@ -1,4 +1,4 @@
-package com.anonymous.coffin.features.base
+package com.anonymous.appilogue.features.base
 
 import android.os.Bundle
 import android.view.View
