+++ app/src/main/java/com/anonymous/appilogue/features/community/CommunityFragment.kt
@@ -1,8 +1,8 @@
-package com.anonymous.coffin.features.community
+package com.anonymous.appilogue.features.community
 
-import com.anonymous.coffin.R
-import com.anonymous.coffin.databinding.FragmentCommunityBinding
-import com.anonymous.coffin.features.base.BaseFragment
+import com.anonymous.appilogue.R
+import com.anonymous.appilogue.databinding.FragmentCommunityBinding
+import com.anonymous.appilogue.features.base.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
