+++ app/src/main/java/com/anonymous/appilogue/features/home/HomeFragment.kt
@@ -1,0 +1,32 @@
+package com.anonymous.appilogue.features.home
+
+import android.os.Bundle
+import android.view.View
+import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.core.view.updateLayoutParams
+import androidx.fragment.app.activityViewModels
+import com.anonymous.appilogue.R
+import com.anonymous.appilogue.databinding.FragmentHomeBinding
+import com.anonymous.appilogue.features.base.BaseFragment
+import com.anonymous.appilogue.features.main.MainActivity
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.google.android.material.bottomsheet.BottomSheetBehavior
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class HomeFragment :
+    BaseFragment<FragmentHomeBinding, HomeViewModel>(R.layout.fragment_home) {
+    override val viewModel: HomeViewModel by activityViewModels()
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.ivMainPlanet.setOnClickListener {
+            (activity as MainActivity).run {
+                findViewById<BottomNavigationView>(R.id.bottomNavigationView).visibility = View.GONE
+                findViewById<ConstraintLayout>(R.id.bottomSheetHome).updateLayoutParams {
+                    height = 600
+                }
+            }
+        }
+    }
+}
