+++ app/src/main/java/com/anonymous/appilogue/features/main/MainActivity.kt
@@ -1,15 +1,20 @@
-package com.anonymous.coffin.features.main
+package com.anonymous.appilogue.features.main
 
 import android.os.Bundle
 import android.view.MenuItem
+import android.view.View
+import androidx.activity.viewModels
+import androidx.core.view.updateLayoutParams
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
-import com.anonymous.coffin.R
-import com.anonymous.coffin.databinding.ActivityMainBinding
-import com.anonymous.coffin.features.base.BaseActivity
-import com.anonymous.coffin.features.community.CommunityFragment
-import com.anonymous.coffin.features.home.HomeFragment
-import com.anonymous.coffin.features.profile.ProfileFragment
+import com.anonymous.appilogue.R
+import com.anonymous.appilogue.databinding.ActivityMainBinding
+import com.anonymous.appilogue.features.base.BaseActivity
+import com.anonymous.appilogue.features.community.CommunityFragment
+import com.anonymous.appilogue.features.home.HomeFragment
+import com.anonymous.appilogue.features.home.HomeViewModel
+import com.anonymous.appilogue.features.profile.ProfileFragment
+import com.google.android.material.bottomsheet.BottomSheetBehavior
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -19,6 +24,8 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     private val profileFragment by lazy { ProfileFragment() }
     private val communityFragment by lazy { CommunityFragment() }
 
+    private val viewModel by viewModels<HomeViewModel>()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         bind {
@@ -26,8 +33,10 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
                 navigateFragment(it)
                 true
             }
+            homeViewModel = viewModel
         }
         setFragment(homeFragment)
+        initBottomSheet()
     }
 
     private fun navigateFragment(menuItem: MenuItem) {
@@ -51,6 +60,28 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
                 show(fragment)
             }
             setPrimaryNavigationFragment(fragment)
+        }
+    }
+
+
+    private fun initBottomSheet() {
+        setBottomSheetHeight()
+        BottomSheetBehavior.from(binding.bottomSheetHome.root)
+            .addBottomSheetCallback(
+                object : BottomSheetBehavior.BottomSheetCallback() {
+                    override fun onStateChanged(bottomSheet: View, newState: Int) {
+                        // TODO
+                    }
+
+                    override fun onSlide(bottomSheet: View, slideOffset: Float) {
+                        // TODO
+                    }
+                })
+    }
+
+    private fun setBottomSheetHeight() {
+        binding.bottomSheetHome.root.updateLayoutParams {
+            // TODO
         }
     }
 }
