--- app/src/main/res/drawable/bg_corner_stroke.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="@dimen/radius_corner" />
-    <stroke
-        android:width="2dp"
-        android:color="@color/gray" />
-</shape>
