+++ app/src/main/res/drawable/bg_home_bottomsheet.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:bottom="-3dp"
+        android:left="1dp"
+        android:right="1dp"
+        android:top="2dp">
+        <shape>
+            <corners
+                android:topLeftRadius="@dimen/radius_corner"
+                android:topRightRadius="@dimen/radius_corner" />
+            <stroke
+                android:width="2dp"
+                android:color="@color/gray" />
+            <solid android:color="@color/half_alpha_black" />
+        </shape>
+    </item>
+</layer-list>
