+++ app/src/main/res/layout/fragment_home.xml
@@ -1,18 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
 
         <variable
             name="homeViewModel"
-            type="com.anonymous.coffin.features.home.HomeViewModel" />
+            type="com.anonymous.appilogue.features.home.HomeViewModel" />
     </data>
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
@@ -30,13 +25,6 @@
                 app:srcCompat="@drawable/star" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <include
-            android:id="@+id/bottomSheetHome"
-            layout="@layout/bottomsheet_home"
-            app:homeViewModel="@{homeViewModel}" />
-
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
 
 
 </layout>
