+++ app/src/main/res/values/colors.xml
@@ -9,4 +9,5 @@
     <color name="white">#FFFFFFFF</color>
     <color name="light_gray">#f5f5f5</color>
     <color name="gray">#2E2E39</color>
+    <color name="half_alpha_black">#80000000</color>
 </resources>
