+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Coffin</string>
+    <string name="app_name">Appilogue</string>
     <string name="menu_home">Home</string>
     <string name="menu_profile">profile</string>
     <string name="menu_community">community</string>
